requirements.txt
fastapi
uvicorn
python-dotenv
motor
passlib[bcrypt]
python-jose
pydantic
pydantic-settings
playwright
beautifulsoup4
urllib3
regex
typing-extensions
fastapi[all]
python-multipart